Refactor Maven's POM
The previous configuration was using deprecated syntax. This commit
fixes that and enables more configuration options for the used plugins.

Some options might be too strict (especially fail as soon as possible,
even on the smallest issues) but we will see in due time
how (un)necessary and annoying this will be.