Add DoubleFileExtensionFilter
The standard FileNameExtensionFilter is not sufficient for our needs, as it filters
only on the last '.' of the file name while our file extensions contain two '.'. We can
work around this issue by filtering on ".graph", but then we would face the issue of
determining which file the user openend first (node or edge). To prevent all this,
here is an implementation of the `FileFilter` interface that filters on double file extensions.